// src/generated/ledger-v1.d.ts
export interface components {
  schemas: {
// ...
    AssetRate: {
      /**
       * Format: date-time
       * @example 2021-01-01T00:00:00Z
       */
      createdAt: string;
      /**
       * Format: uuid
       * @example 00000000-0000-0000-0000-000000000000
       */
      externalId: string;
      /** @example USD */
      from: string;
      /**
       * Format: uuid
       * @example 00000000-0000-0000-0000-000000000000
       */
      id: string;
      /**
       * Format: uuid
       * @example 00000000-0000-0000-0000-000000000000
       */
      ledgerId: string;
      metadata: {
        [key: string]: unknown;
      };
      /**
       * Format: uuid
       * @example 00000000-0000-0000-0000-000000000000
       */
      organizationId: string;
      /**
       * Format: double
       * @example 100
       */
      rate: number;
      /**
       * Format: double
       * @example 2
       */
      scale: number | null;
      /** @example External System */
      source: string | null;
      /** @example BRL */
      to: string;
      /**
       * Format: int64
       * @example 3600
       */
      ttl: number;
      /**
       * Format: date-time
       * @example 2021-01-01T00:00:00Z
       */
      updatedAt: string;
    };
// ...
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}